+++ package.json
@@ -14,6 +14,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "eslint": "8.29.0",
     "eslint-config-next": "13.0.6",
+    "framer-motion": "^10.12.4",
     "maplibre-gl": "^2.4.0",
     "next": "13.0.6",
     "react": "18.2.0",
